テストコード内で書き留めていた具象クラスを別ファイルに移動
