+++ Compiler/Interactors.Tests/Analysis/Obfuscators/VariableDeclarationTest.cs
@@ -1,16 +1,10 @@
-using System;
-using System.Collections.Generic;
 using System.Text;
 
 using KSPCompiler.Domain.Ast.Nodes;
 using KSPCompiler.Domain.Ast.Nodes.Expressions;
-using KSPCompiler.Domain.Ast.Nodes.Extensions;
 using KSPCompiler.Domain.Ast.Nodes.Statements;
-using KSPCompiler.Domain.Symbols;
 using KSPCompiler.Domain.Symbols.MetaData;
-using KSPCompiler.Domain.Symbols.MetaData.Extensions;
-using KSPCompiler.Interactors.Analysis.Obfuscators.Extensions;
-using KSPCompiler.UseCases.Analysis.Evaluations.Declarations;
+using KSPCompiler.Interactors.Analysis.Obfuscators;
 
 using NUnit.Framework;
 
@@ -30,7 +24,7 @@ public class VariableDeclarationTest
 
         symbolTable.Variables.Add( MockUtility.CreateIntVariable( variableName ) );
 
-        var obfuscatedTable = new ObfuscatedVariableTable( symbolTable.Variables, "v" );
+        var obfuscatedTable = new ObfuscatedVariableSymbolTable( symbolTable.Variables, "v" );
 
         var node = new AstVariableDeclarationNode
         {
@@ -61,7 +55,7 @@ public class VariableDeclarationTest
 
         symbolTable.Variables.Add( variable );
 
-        var obfuscatedTable = new ObfuscatedVariableTable( symbolTable.Variables, "v" );
+        var obfuscatedTable = new ObfuscatedVariableSymbolTable( symbolTable.Variables, "v" );
 
         var node = new AstVariableDeclarationNode
         {
@@ -91,7 +85,7 @@ public class VariableDeclarationTest
 
         symbolTable.Variables.Add( variable );
 
-        var obfuscatedTable = new ObfuscatedVariableTable( symbolTable.Variables, "v" );
+        var obfuscatedTable = new ObfuscatedVariableSymbolTable( symbolTable.Variables, "v" );
 
         var node = new AstVariableDeclarationNode
         {
@@ -121,7 +115,7 @@ public class VariableDeclarationTest
 
         symbolTable.Variables.Add( MockUtility.CreateIntVariable( variableName ) );
 
-        var obfuscatedTable = new ObfuscatedVariableTable( symbolTable.Variables, "v" );
+        var obfuscatedTable = new ObfuscatedVariableSymbolTable( symbolTable.Variables, "v" );
 
         var node = new AstVariableDeclarationNode
         {
@@ -160,7 +154,7 @@ public class VariableDeclarationTest
 
         symbolTable.Variables.Add( variable );
 
-        var obfuscatedTable = new ObfuscatedVariableTable( symbolTable.Variables, "v" );
+        var obfuscatedTable = new ObfuscatedVariableSymbolTable( symbolTable.Variables, "v" );
 
         var initializer = new AstArrayInitializerNode
         {
@@ -209,7 +203,7 @@ public class VariableDeclarationTest
 
         symbolTable.Variables.Add( variable );
 
-        var obfuscatedTable = new ObfuscatedVariableTable( symbolTable.Variables, "v" );
+        var obfuscatedTable = new ObfuscatedVariableSymbolTable( symbolTable.Variables, "v" );
 
         var uiInitializer = new AstExpressionListNode();
         uiInitializer.Expressions.Add( new AstIntLiteralNode( 1 ) );
@@ -258,7 +252,7 @@ public class VariableDeclarationTest
 
         symbolTable.Variables.Add( variable );
 
-        var obfuscatedTable = new ObfuscatedVariableTable( symbolTable.Variables, "v" );
+        var obfuscatedTable = new ObfuscatedVariableSymbolTable( symbolTable.Variables, "v" );
 
         var uiInitializer = new AstExpressionListNode();
         uiInitializer.Expressions.Add( new AstIntLiteralNode( 1 ) );
@@ -291,274 +285,4 @@ public class VariableDeclarationTest
 
         Assert.AreEqual( expected, output.ToString() );
     }
-}
-
-public class MockVariableDeclarationVisitor : DefaultAstVisitor
-{
-    private StringBuilder OutputBuilder { get; }
-    private IVariableDeclarationEvaluator Evaluator { get; set; } = new MockVariableDeclarationEvaluator();
-
-    public MockVariableDeclarationVisitor( StringBuilder outputBuilder )
-    {
-        OutputBuilder = outputBuilder;
-    }
-
-    public void Inject( IVariableDeclarationEvaluator evaluator )
-    {
-        Evaluator = evaluator;
-    }
-
-    public override IAstNode Visit( AstVariableDeclarationNode node )
-        => Evaluator.Evaluate( this, node );
-
-    public override IAstNode Visit( AstIntLiteralNode node )
-    {
-        OutputBuilder.Append( node.Value );
-        return node;
-    }
-
-    private class MockVariableDeclarationEvaluator : IVariableDeclarationEvaluator
-    {
-        public IAstNode Evaluate( IAstVisitor visitor, AstVariableDeclarationNode node )
-            => throw new NotImplementedException();
-    }
-}
-
-
-public class VariableDeclarationEvaluator : IVariableDeclarationEvaluator
-{
-    private StringBuilder OutputBuilder { get; }
-    private IVariableSymbolTable Variables { get; }
-    private ISymbolTable<UITypeSymbol> UITypeSymbols { get; }
-    private IObfuscatedVariableTable ObfuscatedTable { get; }
-
-    public VariableDeclarationEvaluator(
-        StringBuilder outputBuilder,
-        IVariableSymbolTable variables,
-        ISymbolTable<UITypeSymbol> uiTypeSymbols,
-        IObfuscatedVariableTable obfuscatedTable )
-    {
-        OutputBuilder   = outputBuilder;
-        Variables       = variables;
-        UITypeSymbols   = uiTypeSymbols;
-        ObfuscatedTable = obfuscatedTable;
-    }
-
-    public IAstNode Evaluate( IAstVisitor visitor, AstVariableDeclarationNode node )
-    {
-        var name = ObfuscatedTable.GetObfuscatedByName( node.Name );
-
-        if( !Variables.TrySearchByName( node.Name, out var variable ) )
-        {
-            throw new KeyNotFoundException( $"Variable not found: {node.Name} from variable symbol table" );
-        }
-
-        OutputBuilder.Append( "declare" );
-
-        foreach( var modifier in node.Modifier.Values )
-        {
-            OutputBuilder.Append( $" {modifier}" );
-        }
-
-        if( variable.UIType != UITypeSymbol.Null )
-        {
-            OutputBuilder.Append( $" {variable.UIType.Name}" );
-        }
-
-        OutputBuilder.Append( $" {name}" );
-
-        if( node.Initializer.IsNotNull() )
-        {
-            OutputInitializer( visitor, node, variable );
-        }
-
-        return node;
-    }
-
-    private void OutputInitializer( IAstVisitor visitor, AstVariableDeclarationNode node, VariableSymbol variable )
-    {
-        if( variable.Modifier.IsUI() )
-        {
-            OutputUIInitializer( visitor, node, variable );
-            return;
-        }
-        else if( variable.DataType.IsArray() )
-        {
-            OutputArrayInitializer( visitor, node );
-            return;
-        }
-
-        OutputPrimitiveInitializer( visitor, node );
-    }
-
-    #region Primitive Initializer
-
-    private void OutputPrimitiveInitializer( IAstVisitor visitor, AstVariableDeclarationNode node )
-    {
-        OutputBuilder.Append( " := " );
-        node.Initializer.PrimitiveInitializer.Accept( visitor );
-    }
-
-    #endregion ~Primitive Initializer
-
-
-    #region Array Initializer
-
-    private void OutputArrayInitializer( IAstVisitor visitor, AstVariableDeclarationNode node )
-    {
-        OutputArraySize( visitor, node.Initializer.ArrayInitializer );
-        OutputArrayElements( visitor, node.Initializer.ArrayInitializer );
-    }
-
-    private void OutputArraySize( IAstVisitor visitor, AstArrayInitializerNode initializer )
-    {
-        OutputBuilder.Append( '[' );
-        initializer.Size.Accept( visitor );
-        OutputBuilder.Append( ']' );
-    }
-
-    private void OutputArrayElements( IAstVisitor visitor, AstArrayInitializerNode initializer )
-    {
-        OutputBuilder.Append( " := " );
-        OutputBuilder.Append( '(' );
-
-        OutputBuilder.AppendExpressionList( visitor, initializer.Initializer );
-
-        OutputBuilder.Append( ')' );
-    }
-
-    #endregion ~Array Initializer
-
-    #region UI Initializer
-
-    private void OutputUIInitializer( IAstVisitor visitor, AstVariableDeclarationNode node, VariableSymbol variable )
-    {
-        if( variable.DataType.IsArray() )
-        {
-            OutputArrayBasedUIInitializer( visitor, node );
-        }
-        else
-        {
-            OutputPrimitiveBasedUIInitializer( visitor, node.Initializer.PrimitiveInitializer.UIInitializer );
-        }
-    }
-
-    private void OutputArrayBasedUIInitializer( IAstVisitor visitor, AstVariableDeclarationNode node )
-    {
-        OutputArraySize( visitor, node.Initializer.ArrayInitializer );
-        OutputUIArguments( visitor, node.Initializer.PrimitiveInitializer.UIInitializer );
-    }
-
-    private void OutputPrimitiveBasedUIInitializer( IAstVisitor visitor, AstExpressionListNode initializer )
-    {
-        OutputUIArguments( visitor, initializer );
-    }
-
-    private void OutputUIArguments( IAstVisitor visitor, AstExpressionListNode expressionList )
-    {
-        OutputBuilder.Append( " := " );
-        OutputBuilder.Append( '(' );
-
-        OutputBuilder.AppendExpressionList( visitor, expressionList );
-
-        OutputBuilder.Append( ')' );
-    }
-
-    #endregion ~UI Initializer
-}
-
-
-
-
-public interface IObfuscatedTable<in TSymbol> where TSymbol : SymbolBase
-{
-    bool TryGetObfuscatedByName( string original, out string result );
-    string GetObfuscatedByName( string original );
-}
-
-public interface IObfuscatedVariableTable : IObfuscatedTable<VariableSymbol> {}
-public interface IObfuscatedUserFunctionTable : IObfuscatedTable<UserFunctionSymbol> {}
-
-public interface IObfuscateFormatter
-{
-    string Format( string original, string prefix, UniqueSymbolIndex index );
-}
-
-public sealed class DefaultObfuscateFormatter : IObfuscateFormatter
-{
-    public string Format( string original, string prefix, UniqueSymbolIndex index )
-        => $"{prefix}{index.Value}";
-}
-
-abstract public class ObfuscatedSymbolTable<TSymbol> : IObfuscatedTable<TSymbol> where TSymbol : SymbolBase
-{
-    // <original name, obfuscated name>
-    private Dictionary<string, string> ObfuscatedTable { get; } = new();
-
-    private string Prefix { get; }
-
-    private IObfuscateFormatter Formatter { get; }
-
-    protected ObfuscatedSymbolTable( ISymbolTable<TSymbol> source, string prefix )
-        : this( source, prefix, new DefaultObfuscateFormatter() ) {}
-
-    protected ObfuscatedSymbolTable( ISymbolTable<TSymbol> source, string prefix, IObfuscateFormatter formatter )
-    {
-        Prefix    = prefix;
-        Formatter = formatter;
-
-        var generator = new UniqueSymbolIndexGenerator();
-        Obfuscate( source, generator );
-    }
-
-    private void Obfuscate( ISymbolTable<TSymbol> source, UniqueSymbolIndexGenerator generator )
-    {
-        foreach( var (name, symbol) in source.Table )
-        {
-            Obfuscate( name, symbol, generator );
-        }
-    }
-
-    private void Obfuscate( string name, TSymbol symbol, UniqueSymbolIndexGenerator generator )
-    {
-        var typePrefix = string.Empty;
-
-        if( symbol.BuiltIn )
-        {
-            return;
-        }
-
-        if( KspRegExpConstants.TypePrefix.IsMatch( name ) )
-        {
-            typePrefix = name[ 0 ].ToString();
-        }
-
-        var obfuscatedName = $"{typePrefix}{Formatter.Format( name, Prefix, generator.Next() )}";
-        ObfuscatedTable.TryAdd( name, obfuscatedName );
-    }
-
-    public bool TryGetObfuscatedByName( string original, out string result )
-    {
-        result = string.Empty;
-
-        if( !ObfuscatedTable.TryGetValue( original, out var obfuscated ) )
-        {
-            return false;
-        }
-
-        result = obfuscated;
-        return true;
-    }
-
-    public string GetObfuscatedByName( string original )
-        => !TryGetObfuscatedByName( original, out var result ) ? original : result;
-}
-
-public class ObfuscatedVariableTable : ObfuscatedSymbolTable<VariableSymbol>, IObfuscatedVariableTable
-{
-    public ObfuscatedVariableTable( IVariableSymbolTable source, string prefix )
-        : base( source, prefix ) {}
-
-    public ObfuscatedVariableTable( IVariableSymbolTable source, string prefix, IObfuscateFormatter formatter )
-        : base( source, prefix, formatter ) {}
 }
