+++ Compiler/Interactors/Analysis/Obfuscators/DefaultObfuscateFormatter.cs
@@ -1,0 +1,10 @@
+using KSPCompiler.Domain.Symbols;
+using KSPCompiler.UseCases.Analysis.Obfuscators;
+
+namespace KSPCompiler.Interactors.Analysis.Obfuscators;
+
+public sealed class DefaultObfuscateFormatter : IObfuscateFormatter
+{
+    public string Format( string original, string prefix, UniqueSymbolIndex index )
+        => $"{prefix}{index.Value}";
+}
