+++ Compiler/UseCases/Analysis/Obfuscators/IObfuscateFormatter.cs
@@ -1,0 +1,8 @@
+using KSPCompiler.Domain.Symbols;
+
+namespace KSPCompiler.UseCases.Analysis.Obfuscators;
+
+public interface IObfuscateFormatter
+{
+    string Format( string original, string prefix, UniqueSymbolIndex index );
+}
